Toolbelt, safety config
+Added the toolbelt, the technomancy armor pants. Dying while wearing the toolbelt will cause the player to respawn with their hotbar and offhand intact. Basically useless if keepInventory is enabled
+Added a 'safe mode' config for temporal entropy, where machines that overfill with temporal entropy will not break or cause any damage, but will shut down until the entropy is handled.

+++ src/main/java/com/Da_Technomancer/crossroads/API/technomancy/FluxUtil.java
@@ -117,15 +117,18 @@ public class FluxUtil{
 	/**
 	 * Checks whether flux is over the limit, and if so performs a flux event and destroys this block
 	 * @param te The machine to check
-	 * @return Whether this machine was destroyed (flux > fluxLimit)
+	 * @return Whether this machine should shut down (flux > fluxLimit && safe mode enabled)
 	 */
 	public static boolean checkFluxOverload(IFluxLink te){
 		if(te.getFlux() > te.getMaxFlux()){
-			World world = te.getTE().getWorld();
-			BlockPos pos = te.getTE().getPos();
+			if(CRConfig.fluxSafeMode.get()){
+				return true;
+			}
+			TileEntity tileEntity = te.getTE();
+			World world = tileEntity.getWorld();
+			BlockPos pos = tileEntity.getPos();
 			world.destroyBlock(pos, CRConfig.entropyDropBlock.get());
 			fluxEvent(world, pos);
-			return true;
 		}
 		return false;
 	}
